LayoutTests/inspector-protocol/layer-tree.html

<html>
<head>
<script type="text/javascript" src="../http/tests/inspector-protocol/resources/protocol-test.js"></script>
<script type="text/javascript">

function addCompositedLayer()
{
    var element = document.createElement("div");
    element.className = "composited";
    element.id = "last-element";
    document.body.appendChild(element);
};

function test()
{
    var initialLayerTree;
    var eventsCount = 0;

    InspectorTest.eventHandler["LayerTree.layerTreeDidChange"] = function (messageObject) {
        eventsCount++;
    };

    getDocument();
    
    function getDocument()
    {
        // We must first get the document so that later on we may get sensible nodeIds.
        step({
            name: "Get the Document",
            command: "DOM.getDocument",
            parameters: {},
            callback: enableLayerTreeAgent
        });
    };
    
    function enableLayerTreeAgent(result)
    {
        step({
            name: "Enable the LayerTree agent",
            command: "LayerTree.enable",
            parameters: {},
            callback: getInitialLayerTree
        });
    };

    function getInitialLayerTree(result)
    {
        step({
            name: "Get the initial layer tree",
            command: "LayerTree.getLayerTree",
            parameters: {},
            callback: gotInitialLayerTree
        });
    };

    function gotInitialLayerTree(result)
    {
        initialLayerTree = result.layerTree;

        dumpLayerTree(initialLayerTree);

        step({
            name: "Message the page to add a new composited layer",
            command: "Runtime.evaluate",
            parameters: {"expression": "addCompositedLayer()"},
            callback: getModifiedLayerTree
        });
    };

    function getModifiedLayerTree(result)
    {
        step({
            name: "Get the modified layer tree",
            command: "LayerTree.getLayerTree",
            parameters: {},
            callback: gotModifiedLayerTree
        });
    };

    var layerCount = 0;

    function gotModifiedLayerTree(result)
    {
        dumpLayerTree(result.layerTree);

        var mutations = layerTreeMutations(initialLayerTree, result.layerTree);
        var lastId = mutations.additions[0];

        step({
            name: "Obtain the node id for the newly inserted layer",
            command: "LayerTree.nodeIdForLayerId",
            parameters: {"layerId": lastId},
            callback: gotNodeIdForLastLayer
        });
    };

    function gotNodeIdForLastLayer(result)
    {
        var id = result.nodeId;
        step({
            name: "Get attributes for the newly inserted node",
            command: "DOM.getAttributes",
            parameters: {"nodeId": id},
            callback: gotNodeAttributes
        });
    };

    function gotNodeAttributes(result)
    {
        var attributes = attributesDictionaryFromArray(result.attributes);
        if (attributes.id !== "last-element")
            InspectorTest.log("FAIL: Did not obtain the expected element for the last inserted layer.");

        finishTest();
    };

    function finishTest()
    {
        if (!eventsCount)
            InspectorTest.log("FAIL: Did not receive layerTreeDidChange events.");
        else
            InspectorTest.log("\n=== Test complete, all expected conditions met ===");

        InspectorTest.completeTest();
    };

    function layerTreeMutations(oldLayerTree, newLayerTree)
    {
        var oldKeys = Object.keys(flattenedLayerTree(oldLayerTree));
        var newKeys = Object.keys(flattenedLayerTree(newLayerTree));

        return {
            additions: newKeys.filter(function (key) {
                return (oldKeys.indexOf(key) === -1);
            }),
            removals: oldKeys.filter(function (key) {
                return (newKeys.indexOf(key) === -1);
            })
        };
    };

    function flattenedLayerTree(layerTree)
    {
        var layerByIds = {};

        function recurse(layer)
        {
            layerByIds[layer.layerId] = layer;
            if (layer.childLayers)
                layer.childLayers.forEach(recurse);
        };
        
        recurse(layerTree);

        return layerByIds;
    };

    function attributesDictionaryFromArray(attributes)
    {
        var dictionary = {}
        for (var i = 0, count = attributes.length; i < count; i += 2) {
            dictionary[attributes[i]] = attributes[i + 1];
        }
        return dictionary;
    };

    function dumpLayerTree(layerTree)
    {
        function replacer(key, value)
        {
            if (key === "layerId" || key === "memory" || key === "paintCount")
                return typeof(value);

            // some values differ based on port, but the ones we most
            // care about will always be less or equal 100.
            if ((key === "width" || key === "height") && value > 100) 
                return typeof(value);
            
            return value;
        };
        
        InspectorTest.log("\n" + JSON.stringify(layerTree, replacer, "    "));
    };

    function step(test)
    {
        InspectorTest.log("\n=== " + test.name + " ===\n")
        InspectorTest.sendCommand(test.command, test.parameters, function(messageObject) {
            if (messageObject.hasOwnProperty("error")) {
                InspectorTest.log("FAIL: " + messageObject.error.message + " (" + messageObject.error.code + ")");
                InspectorTest.completeTest();
                return;
            }

            InspectorTest.log("PASS");
            test.callback(messageObject.result);
        });
    };

};

window.addEventListener("DOMContentLoaded", function () {
    runTest();
}, false);

</script>
<style type="text/css">
      
    div {
        position: absolute;
        top: 0;
        left: 0;
    }
      
    .regular {
        width: 100px;
        height: 100px;
        background-color: black;
    }

    .composited {
        top: 25px;
        left: 25px;
        width: 50px;
        height: 50px;
        background-color: blue;
        -webkit-transform: translateZ(0);
    }
      
    .offset {
        left: 200px;
        -webkit-transform: translateZ(0);
    }

</style>
</head>
<body>

    <div class="regular"></div>

    <div class="composited">
        <div class="composited"></div>
    </div>

    <div class="regular offset">
        <div class="composited"></div>
    </div>

</body>
</html>
