LayoutTests/inspector-protocol/layer-tree-expected.txt


=== Get the Document ===

PASS

=== Enable the LayerTree agent ===

PASS

=== Get the initial layer tree ===

PASS

{
    "layerId": "string",
    "bounds": {
        "x": 0,
        "y": 0,
        "width": "number",
        "height": "number"
    },
    "isComposited": true,
    "memory": "number",
    "compositedBounds": {
        "x": 0,
        "y": 0,
        "width": "number",
        "height": "number"
    },
    "paintCount": "number",
    "childLayers": [
        {
            "layerId": "string",
            "bounds": {
                "x": 0,
                "y": 0,
                "width": "number",
                "height": "number"
            },
            "isComposited": false,
            "childLayers": [
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": 100,
                        "height": 100
                    },
                    "isComposited": false,
                    "childLayers": []
                },
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": 50,
                        "height": 50
                    },
                    "isComposited": true,
                    "memory": "number",
                    "compositedBounds": {
                        "x": 0,
                        "y": 0,
                        "width": 50,
                        "height": 50
                    },
                    "paintCount": "number",
                    "childLayers": [
                        {
                            "layerId": "string",
                            "bounds": {
                                "x": 0,
                                "y": 0,
                                "width": 50,
                                "height": 50
                            },
                            "isComposited": true,
                            "memory": "number",
                            "compositedBounds": {
                                "x": 0,
                                "y": 0,
                                "width": 50,
                                "height": 50
                            },
                            "paintCount": "number",
                            "childLayers": []
                        }
                    ]
                },
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": 100,
                        "height": 100
                    },
                    "isComposited": true,
                    "memory": "number",
                    "compositedBounds": {
                        "x": 0,
                        "y": 0,
                        "width": 100,
                        "height": 100
                    },
                    "paintCount": "number",
                    "childLayers": [
                        {
                            "layerId": "string",
                            "bounds": {
                                "x": 0,
                                "y": 0,
                                "width": 50,
                                "height": 50
                            },
                            "isComposited": true,
                            "memory": "number",
                            "compositedBounds": {
                                "x": 0,
                                "y": 0,
                                "width": 50,
                                "height": 50
                            },
                            "paintCount": "number",
                            "childLayers": []
                        }
                    ]
                },
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": 0,
                        "height": 0
                    },
                    "isComposited": false,
                    "childLayers": []
                },
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": 0,
                        "height": 0
                    },
                    "isComposited": false,
                    "childLayers": []
                },
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": "number",
                        "height": "number"
                    },
                    "isComposited": true,
                    "memory": "number",
                    "compositedBounds": {
                        "x": 0,
                        "y": 0,
                        "width": "number",
                        "height": "number"
                    },
                    "paintCount": "number",
                    "childLayers": []
                }
            ]
        }
    ]
}

=== Message the page to add a new composited layer ===

PASS

=== Get the modified layer tree ===

PASS

{
    "layerId": "string",
    "bounds": {
        "x": 0,
        "y": 0,
        "width": "number",
        "height": "number"
    },
    "isComposited": true,
    "memory": "number",
    "compositedBounds": {
        "x": 0,
        "y": 0,
        "width": "number",
        "height": "number"
    },
    "paintCount": "number",
    "childLayers": [
        {
            "layerId": "string",
            "bounds": {
                "x": 0,
                "y": 0,
                "width": "number",
                "height": "number"
            },
            "isComposited": false,
            "childLayers": [
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": 100,
                        "height": 100
                    },
                    "isComposited": false,
                    "childLayers": []
                },
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": 50,
                        "height": 50
                    },
                    "isComposited": true,
                    "memory": "number",
                    "compositedBounds": {
                        "x": 0,
                        "y": 0,
                        "width": 50,
                        "height": 50
                    },
                    "paintCount": "number",
                    "childLayers": [
                        {
                            "layerId": "string",
                            "bounds": {
                                "x": 0,
                                "y": 0,
                                "width": 50,
                                "height": 50
                            },
                            "isComposited": true,
                            "memory": "number",
                            "compositedBounds": {
                                "x": 0,
                                "y": 0,
                                "width": 50,
                                "height": 50
                            },
                            "paintCount": "number",
                            "childLayers": []
                        }
                    ]
                },
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": 100,
                        "height": 100
                    },
                    "isComposited": true,
                    "memory": "number",
                    "compositedBounds": {
                        "x": 0,
                        "y": 0,
                        "width": 100,
                        "height": 100
                    },
                    "paintCount": "number",
                    "childLayers": [
                        {
                            "layerId": "string",
                            "bounds": {
                                "x": 0,
                                "y": 0,
                                "width": 50,
                                "height": 50
                            },
                            "isComposited": true,
                            "memory": "number",
                            "compositedBounds": {
                                "x": 0,
                                "y": 0,
                                "width": 50,
                                "height": 50
                            },
                            "paintCount": "number",
                            "childLayers": []
                        }
                    ]
                },
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": 0,
                        "height": 0
                    },
                    "isComposited": false,
                    "childLayers": []
                },
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": 0,
                        "height": 0
                    },
                    "isComposited": false,
                    "childLayers": []
                },
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": "number",
                        "height": "number"
                    },
                    "isComposited": true,
                    "memory": "number",
                    "compositedBounds": {
                        "x": 0,
                        "y": 0,
                        "width": "number",
                        "height": "number"
                    },
                    "paintCount": "number",
                    "childLayers": []
                },
                {
                    "layerId": "string",
                    "bounds": {
                        "x": 0,
                        "y": 0,
                        "width": 50,
                        "height": 50
                    },
                    "isComposited": true,
                    "memory": "number",
                    "compositedBounds": {
                        "x": 0,
                        "y": 0,
                        "width": 50,
                        "height": 50
                    },
                    "paintCount": "number",
                    "childLayers": []
                }
            ]
        }
    ]
}

=== Obtain the node id for the newly inserted layer ===

PASS

=== Get attributes for the newly inserted node ===

PASS

=== Test complete, all expected conditions met ===
